+++ changelog.md
@@ -1,2 +1,2 @@
-Doing It Myself 1.0.1
-- Fix FoodDataMixin priority
+Doing It Myself 1.0.2
+- Fix a few stack size issues with Panacea